+++ exports/wait-for-ammo-instantiation.ts
@@ -25,4 +25,4 @@
 
 import '../cocos/physics/ammo/polyfill-atob';
 
-export { waitForAmmoInstantiation as default } from '../cocos/physics/ammo/ammo-instantiated';
+export { waitForAmmoInstantiation as default } from '../cocos/physics/ammo/instantiated';
